+++ src/Constants.php
@@ -18,13 +18,14 @@ class Constants
 
     // API Configuration
     public const API_BASE_URL_DEFAULT = 'https://api.upsnap.ai';
-    public const UPSNAP_DASHBOARD_URL = 'https://upsnap.ai';
+    public const UPSNAP_DASHBOARD_URL = 'https://app.upsnap.ai';
     public const UPSNAP_STATS_PAGE_URL = 'https://stats.upsnap.ai';
     public const API_VERSION = 'v1';
     // API Endpoints
     public const ENDPOINT_HEALTHCHECK = 'healthcheck';
 
     public const ENDPOINT_VERIFY_API_KEY = 'tokens/validate';
+    public const REGISTER_SOURCE = 'craftcms';
 
     // Check Types
     public const CHECK_REACHABILITY = 'reachability';
@@ -153,6 +154,7 @@ class Constants
         'active' => 'active',
         'suspended' => 'suspended',
         'expired' => 'expired',
+        'account_expired' => 'account_expired',
         'deleted' => 'deleted',
     ];
 
@@ -188,14 +190,25 @@ class Constants
                 'create' => 'user/status-pages',
                 'update' => 'user/status-pages',
                 'delete' => 'user/status-pages',
-            ]
+            ],
+            'monitors_stats' => 'user/monitors/uptime-stats',
         ],
         'user' => [
-            'details' => 'user/details'
+            'details' => 'user/details',
+            'register' => 'user/register',
+            'login' => 'user/login'
         ],
         'regions' => [
             'list' => 'regions'
         ],
+        'tags' => [
+            'list' => 'user/tags',
+            'create' => 'user/tags',
+        ],
+        'tokens' => [
+            'list' => 'tokens',
+            'generate' => 'tokens/generate',
+        ]
     ];
 
     public const LIGHTHOUSE_STRATEGY = [