+++ src/assetbundles/SettingsAsset.php
@@ -13,7 +13,11 @@ class SettingsAsset extends BaseAsset
         $this->js[] = 'js/monitors.js';
         $this->js[] = 'js/monitorsTable.js';
         $this->js[] = 'js/notificationChannels.js';
+        $this->js[] = 'js/signupInlineProgress.js';
+        $this->js[] = 'js/verificationBanner.js';
         $this->css[] = 'css/addMonitorModal.css';
         $this->css[] = 'css/notification-channels.css';
+        $this->css[] = 'css/signupInlineProgress.css';
+        $this->css[] = 'css/verificationBanner.css';
     }
 }